+++ examples/query-api.graphql
@@ -1,15 +1,61 @@
 # Code generated by Compass integration tests, DO NOT EDIT.
 query {
-  result: api(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
-    id
-    name
-    description
-    spec {
-      data
-      format
-      type
-      fetchRequest {
-        url
+  result: application(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+    api(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+      id
+      name
+      description
+      spec {
+        data
+        format
+        type
+        fetchRequest {
+          url
+          auth {
+            credential {
+              ... on BasicCredentialData {
+                username
+                password
+              }
+              ... on OAuthCredentialData {
+                clientId
+                clientSecret
+                url
+              }
+            }
+            additionalHeaders
+            additionalQueryParams
+            requestAuth {
+              csrf {
+                tokenEndpointURL
+                credential {
+                  ... on BasicCredentialData {
+                    username
+                    password
+                  }
+                  ... on OAuthCredentialData {
+                    clientId
+                    clientSecret
+                    url
+                  }
+                }
+                additionalHeaders
+                additionalQueryParams
+              }
+            }
+          }
+          mode
+          filter
+          status {
+            condition
+            timestamp
+          }
+        }
+      }
+      targetURL
+      group
+      auths {
+        runtimeID
         auth {
           credential {
             ... on BasicCredentialData {
@@ -43,19 +89,8 @@ query {
             }
           }
         }
-        mode
-        filter
-        status {
-          condition
-          timestamp
-        }
       }
-    }
-    targetURL
-    group
-    auths {
-      runtimeID
-      auth {
+      defaultAuth {
         credential {
           ... on BasicCredentialData {
             username
@@ -88,45 +123,12 @@ query {
           }
         }
       }
-    }
-    defaultAuth {
-      credential {
-        ... on BasicCredentialData {
-          username
-          password
-        }
-        ... on OAuthCredentialData {
-          clientId
-          clientSecret
-          url
-        }
+      version {
+        value
+        deprecated
+        deprecatedSince
+        forRemoval
       }
-      additionalHeaders
-      additionalQueryParams
-      requestAuth {
-        csrf {
-          tokenEndpointURL
-          credential {
-            ... on BasicCredentialData {
-              username
-              password
-            }
-            ... on OAuthCredentialData {
-              clientId
-              clientSecret
-              url
-            }
-          }
-          additionalHeaders
-          additionalQueryParams
-        }
-      }
-    }
-    version {
-      value
-      deprecated
-      deprecatedSince
-      forRemoval
     }
   }
 }
