+++ examples/query-event-api.graphql
@@ -1,63 +1,65 @@
 # Code generated by Compass integration tests, DO NOT EDIT.
 query {
-  result: eventAPI(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
-    id
-    applicationID
-    name
-    description
-    group
-    spec {
-      data
-      type
-      format
-      fetchRequest {
-        url
-        auth {
-          credential {
-            ... on BasicCredentialData {
-              username
-              password
-            }
-            ... on OAuthCredentialData {
-              clientId
-              clientSecret
-              url
-            }
-          }
-          additionalHeaders
-          additionalQueryParams
-          requestAuth {
-            csrf {
-              tokenEndpointURL
-              credential {
-                ... on BasicCredentialData {
-                  username
-                  password
-                }
-                ... on OAuthCredentialData {
-                  clientId
-                  clientSecret
-                  url
-                }
+  result: application(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+    eventAPI(id: "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa") {
+      id
+      applicationID
+      name
+      description
+      group
+      spec {
+        data
+        type
+        format
+        fetchRequest {
+          url
+          auth {
+            credential {
+              ... on BasicCredentialData {
+                username
+                password
               }
-              additionalHeaders
-              additionalQueryParams
+              ... on OAuthCredentialData {
+                clientId
+                clientSecret
+                url
+              }
+            }
+            additionalHeaders
+            additionalQueryParams
+            requestAuth {
+              csrf {
+                tokenEndpointURL
+                credential {
+                  ... on BasicCredentialData {
+                    username
+                    password
+                  }
+                  ... on OAuthCredentialData {
+                    clientId
+                    clientSecret
+                    url
+                  }
+                }
+                additionalHeaders
+                additionalQueryParams
+              }
             }
           }
-        }
-        mode
-        filter
-        status {
-          condition
-          timestamp
+          mode
+          filter
+          status {
+            condition
+            timestamp
+          }
         }
       }
-    }
-    version {
-      value
-      deprecated
-      deprecatedSince
-      forRemoval
+      version {
+        value
+        deprecated
+        deprecatedSince
+        forRemoval
+      }
     }
   }
 }
